5min 10* bar alert

+++ start.sh
@@ -1,9 +1,22 @@
 #!/bin/bash
-conda activate mytrader
-rm -rf alert_5min_0.log
-rm -rf alert_5min_1.log
-rm -rf alert_5min_2.log
 
-nohup python3 binance_websocket.py 0 > alert_5min_0.log 2>&1 &
-nohup python3 binance_websocket.py 1 > alert_5min_1.log 2>&1 &
-nohup python3 binance_websocket.py 2 > alert_5min_2.log 2>&1 &
+#!/bin/bash
+
+while true;
+do
+    DATE=`date | cut -d' ' -f4`
+    echo $DATE
+    if [[ $DATE == "3:35:00" ]]
+    then
+        conda activate mytrader
+        rm -rf alert_5min_0.log
+        rm -rf alert_5min_1.log
+        rm -rf alert_5min_2.log
+
+        nohup python3 binance_websocket.py 0 > alert_5min_0.log 2>&1 &
+        nohup python3 binance_websocket.py 1 > alert_5min_1.log 2>&1 &
+        nohup python3 binance_websocket.py 2 > alert_5min_2.log 2>&1 &
+
+    fi
+done
+
